set-head: no update without change and better output for --auto
Currently, even if there is no actual change to remote/HEAD calling
remote set-head will overwrite the appropriate file and if set to --auto
will also print a message saying "remote/HEAD set to branch", which
implies something was changed.

Change the behaviour of remote set-head so that the reference is only
updated if it actually needs to change.

Change the output of --auto, so the output actually reflects what was
done: a) set a previously unset HEAD, b) change HEAD because remote
changed or c) no updates. Make the output easily parsable, by using
a slightly clunky wording that allows all three outputs to have the same
structure and number of words.

Signed-off-by: Bence Ferdinandy <bence@ferdinandy.com>