notes: add prepend command
When a note is detailing commit history, it makes sense to keep the
latest change on top, but unlike adding things at the bottom with
"git notes append" this can only be done manually. Add a

    git notes prepend

command, which works exactly like the append command, except that it
inserts the text before the current contents of the note instead of
after.

Signed-off-by: Bence Ferdinandy <bence@ferdinandy.com>